refactor(money): BigDecimal on Product and Cart, and the pricing pipeline in Money
Slice 3 of the money migration, and the one that removes the seams slice 2 left
behind. Orders were exact, but every order total is a sum of cart lines and every
cart line starts from a product price — so an exact order built out of
floating-point inputs was only exact by luck.

- Product.price / discount / specialPrice, Cart.totalPrice and
  CartItem.discount / productPrice are BigDecimal at NUMERIC(12,2) (V25), with
  their DTOs, OrderSummaryDTO, BundleDTO and PriceLine following.
- The pricing pipeline speaks Money end to end now: PricingContext takes it,
  ShippingCalculator takes and returns it (its thresholds are Money constants),
  and ShippingRule no longer converts anything. The four toDouble() calls slice 2
  left in OrderServiceImpl are gone, and so is its money() boundary helper — it
  has no callers left, which is exactly what that commit predicted.
- Two hand-rolled roundings went with them. BundleServiceImpl computed
  Math.round(x * 100.0) / 100.0 three separate times, which could leave savings a
  cent off the difference between the two figures it was derived from; and a
  special price was price - (discount * 0.01 * price), which for 25% off 84.99
  gives 63.742499999999996. Both are Money.percentage now, rounding to the cent
  at each step, so the figures agree with each other by construction.
- Product prices are pinned to the cent on the way in. A JSON body carries
  whatever scale the client wrote — "120", "120.0" — and the entity used to hold
  that until a database round trip normalised it. Normalising at the boundary
  means nothing downstream ends up comparing 120.0 with 120.00.
- Faceted search: ProductFilter's price bounds and the configured band edges are
  BigDecimal, the edges read straight from the configuration text instead of
  through Double.valueOf on their way to a NUMERIC column. bandExpression is
  generic over Number, so the rating bands — which are not money — keep doubles.

Money gains times(int), because a line total is a unit price counted n times, and
negated(), for the sign a discount takes on a printed bill.

Test (Testcontainers): the six columns are numeric, and twenty cart lines at 0.07
sum in the database to exactly 1.40, which a float column could not promise.
Unit: 25% off 84.99 is 63.74 with no remainder, and a price sent as "120" is
stored to the cent.

Still double: Bundle, PromoCampaign, ReturnRequest and SubscriptionPlan — the
last slice. VAT is the other thing outstanding, and it is not a conversion: the
pipeline has a TAX line type and a taxTotal() accumulator, but no rule fills
them, so an order currently carries no tax line at all.

+++ ecom-backend/src/main/java/com/ecommerce/project/controller/ProductController.java
@@ -17,6 +17,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -106,8 +107,8 @@ public class ProductController extends BaseController {
     public ResponseEntity<FacetedProductResponse> searchProductsFaceted(
             @RequestParam(name = "q", required = false) String keyword,
             @RequestParam(name = "categoryId", required = false) List<Long> categoryIds,
-            @RequestParam(name = "minPrice", required = false) Double minPrice,
-            @RequestParam(name = "maxPrice", required = false) Double maxPrice,
+            @RequestParam(name = "minPrice", required = false) BigDecimal minPrice,
+            @RequestParam(name = "maxPrice", required = false) BigDecimal maxPrice,
             @RequestParam(name = "minRating", required = false) Double minRating,
             @RequestParam(name = "inStock", required = false) Boolean inStock,
             @ModelAttribute PaginationParams params) {
