+++ ecom-backend/src/main/java/com/ecommerce/project/service/AdminAuditLogService.java
@@ -2,13 +2,14 @@ package com.ecommerce.project.service;
 
 import com.ecommerce.project.model.AdminAuditLog;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AdminAuditLogService {
 
     AdminAuditLog logPriceChange(Long adminUserId, String adminUsername, Long productId,
-                                 double oldPrice, double newPrice,
-                                 double oldSpecialPrice, double newSpecialPrice);
+                                 BigDecimal oldPrice, BigDecimal newPrice,
+                                 BigDecimal oldSpecialPrice, BigDecimal newSpecialPrice);
 
     AdminAuditLog logRoleChange(Long adminUserId, String adminUsername, Long userId,
                                 String oldRoles, String newRoles);
